Add files via upload
This is an old document (2007, and a draft apparently) that nonetheless describes DIS in an introductory kind of way. I have highlighted anything that I thought may be useful for us. We'll also have to look into the OpenDIS implementation of this protocol though, and look into the Entity State PDUs in more detail.