Notes on Tran's part of the meeting
I jotted down what I understood from what Dr Sura was saying to do, and found a link myself that might be useful in calculating the rotation of the Birds. Hopefully I didn't miss too much of it! Keep in mind I didn't take notes for the rest of the meeting, just this part (as this part went for so long as was so dense)